pyproject.toml
[project]
name = "agnostas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "clip>=0.2.0",
    "dotenv>=0.9.9",
    "e2b>=2.13.0",
    "langchain>=1.2.8",
    "langchain-core>=1.2.8",
    "langchain-groq>=1.1.2",
    "langgraph>=1.0.7",
    "matplotlib>=3.10.8",
    "modal>=1.3.2",
    "numpy>=2.4.2",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.46",
    "streamlit>=1.53.1",
    "torch>=2.10.0",
    "transformers>=5.0.0",
    "ultralytics>=8.4.10",
]
